Moved week controlling to Fragment
