+++ app/src/main/java/org/zephyrsoft/trackworktime/WorkTimeTrackerActivity.java
@@ -41,28 +41,20 @@ import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Spinner;
-import android.widget.TableLayout;
-import android.widget.TableRow;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import org.pmw.tinylog.Logger;
 import org.zephyrsoft.trackworktime.database.DAO;
-import org.zephyrsoft.trackworktime.model.DayLine;
 import org.zephyrsoft.trackworktime.model.Event;
-import org.zephyrsoft.trackworktime.model.PeriodEnum;
 import org.zephyrsoft.trackworktime.model.Task;
-import org.zephyrsoft.trackworktime.model.TimeSum;
 import org.zephyrsoft.trackworktime.model.TypeEnum;
 import org.zephyrsoft.trackworktime.model.Week;
-import org.zephyrsoft.trackworktime.model.WeekDayEnum;
-import org.zephyrsoft.trackworktime.model.WeekPlaceholder;
 import org.zephyrsoft.trackworktime.options.Key;
-import org.zephyrsoft.trackworktime.timer.TimeCalculator;
 import org.zephyrsoft.trackworktime.timer.TimerManager;
 import org.zephyrsoft.trackworktime.util.DateTimeUtil;
 import org.zephyrsoft.trackworktime.util.ExternalNotificationManager;
 import org.zephyrsoft.trackworktime.util.PreferencesUtil;
+import org.zephyrsoft.trackworktime.util.WeekFragment;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -76,14 +68,14 @@ import java.io.OutputStreamWriter;
 import java.util.List;
 import java.util.TimeZone;
 
-import hirondelle.date4j.DateTime;
+import static org.zephyrsoft.trackworktime.util.WeekFragment.WeekCallback;
 
 /**
  * Main activity of the application.
  *
  * @author Mathis Dirksen-Thedens
  */
-public class WorkTimeTrackerActivity extends AppCompatActivity {
+public class WorkTimeTrackerActivity extends AppCompatActivity implements WeekCallback {
 	private static final int PERMISSION_REQUEST_CODE_BACKUP = 1;
 	private static final int PERMISSION_REQUEST_CODE_RESTORE = 2;
 	private static final int PERMISSION_REQUEST_CODE_AUTOMATIC_BACKUP = 3;
@@ -96,68 +88,11 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 		}
 	}
 
-	private TableLayout weekTable = null;
-	private TableRow titleRow = null;
-	private TextView topLeftCorner = null;
-	private TextView inLabel = null;
-	private TextView outLabel = null;
-	private TextView workedLabel = null;
-	private TextView flexiLabel = null;
-	private TableRow mondayRow = null;
-	private TextView mondayLabel = null;
-	private TextView mondayIn = null;
-	private TextView mondayOut = null;
-	private TextView mondayWorked = null;
-	private TextView mondayFlexi = null;
-	private TableRow tuesdayRow = null;
-	private TextView tuesdayLabel = null;
-	private TextView tuesdayIn = null;
-	private TextView tuesdayOut = null;
-	private TextView tuesdayWorked = null;
-	private TextView tuesdayFlexi = null;
-	private TableRow wednesdayRow = null;
-	private TextView wednesdayLabel = null;
-	private TextView wednesdayIn = null;
-	private TextView wednesdayOut = null;
-	private TextView wednesdayWorked = null;
-	private TextView wednesdayFlexi = null;
-	private TableRow thursdayRow = null;
-	private TextView thursdayLabel = null;
-	private TextView thursdayIn = null;
-	private TextView thursdayOut = null;
-	private TextView thursdayWorked = null;
-	private TextView thursdayFlexi = null;
-	private TableRow fridayRow = null;
-	private TextView fridayLabel = null;
-	private TextView fridayIn = null;
-	private TextView fridayOut = null;
-	private TextView fridayWorked = null;
-	private TextView fridayFlexi = null;
-	private TableRow saturdayRow = null;
-	private TextView saturdayLabel = null;
-	private TextView saturdayIn = null;
-	private TextView saturdayOut = null;
-	private TextView saturdayWorked = null;
-	private TextView saturdayFlexi = null;
-	private TableRow sundayRow = null;
-	private TextView sundayLabel = null;
-	private TextView sundayIn = null;
-	private TextView sundayOut = null;
-	private TextView sundayWorked = null;
-	private TextView sundayFlexi = null;
-	private TableRow totalRow = null;
-	private TextView totalLabel = null;
-	private TextView totalIn = null;
-	private TextView totalOut = null;
-	private TextView totalWorked = null;
-	private TextView totalFlexi = null;
 	private Spinner task = null;
 	private EditText text = null;
 	private Button clockInButton = null;
 	private Button clockOutButton = null;
-	private Button previousWeekButton = null;
-	private Button nextWeekButton = null;
-	private Button todayButton = null;
+	private WeekFragment weekFragment = null;
 
 	private static WorkTimeTrackerActivity instance = null;
 
@@ -169,11 +104,9 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 	private DAO dao = null;
 	private TimerManager timerManager = null;
 	private ExternalNotificationManager externalNotificationManager = null;
-	private TimeCalculator timeCalculator = null;
 	private ArrayAdapter<Task> tasksAdapter;
 	private boolean reloadTasksOnResume = false;
 	private List<Task> tasks;
-	private Week currentlyShownWeek;
 
 	private void checkAllOptions() {
 		int disabledSections = PreferencesUtil.checkAllPreferenceSections();
@@ -203,7 +136,6 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 		dao = basics.getDao();
 		timerManager = basics.getTimerManager();
 		externalNotificationManager = basics.getExternalNotificationManager();
-		timeCalculator = basics.getTimeCalculator();
 
 		backupToSdAutomatically();
 
@@ -211,11 +143,7 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 
 		findAllViewsById();
 
-		weekTable.setOnClickListener(v -> showEventList());
-
-		previousWeekButton.setOnClickListener(v -> changeDisplayedWeek(-1));
-
-		nextWeekButton.setOnClickListener(v -> changeDisplayedWeek(1));
+		weekFragment = (WeekFragment)getSupportFragmentManager().findFragmentById(R.id.week_fragment);
 
 		clockInButton.setOnClickListener(v -> clockInAction(0));
 		clockInButton.setOnLongClickListener(v -> {
@@ -245,15 +173,6 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
             return true;
         });
 
-		todayButton.setOnClickListener(v -> {
-            final String todaysWeekStart = DateTimeUtil.getWeekStartAsString(DateTimeUtil.getCurrentDateTime());
-            Week todaysWeek = dao.getWeek(todaysWeekStart);
-            if (todaysWeek == null) {
-                todaysWeek = new WeekPlaceholder(todaysWeekStart);
-            }
-            currentlyShownWeek = todaysWeek;
-            refreshView();
-        });
 
 		// delegate the rest of the work to onResume()
 		reloadTasksOnResume = true;
@@ -339,33 +258,6 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 		refreshView();
 	}
 
-	/**
-	 * @param interval
-	 *            position of the week relative to the currently displayed week, e.g. -2 for two weeks before the
-	 *            currently displayed week
-	 */
-	private void changeDisplayedWeek(int interval) {
-		if (interval == 0) {
-			return;
-		}
-
-		DateTime targetWeekStart = DateTimeUtil.stringToDateTime(currentlyShownWeek.getStart()).plusDays(interval * 7);
-		Week targetWeek = dao.getWeek(DateTimeUtil.dateTimeToString(targetWeekStart));
-		if (targetWeek == null) {
-			// don't insert a new week into the DB but only use a placeholder
-			targetWeek = new WeekPlaceholder(DateTimeUtil.dateTimeToString(targetWeekStart));
-		}
-
-		// display a Toast indicating the change interval (helps the user for more than one week difference)
-		if (Math.abs(interval) > 1) {
-			CharSequence backwardOrForward = interval < 0 ? getText(R.string.backward) : getText(R.string.forward);
-			Toast.makeText(this, backwardOrForward + " " + Math.abs(interval) + " " + getText(R.string.weeks),
-				Toast.LENGTH_SHORT).show();
-		}
-
-		currentlyShownWeek = targetWeek;
-		refreshView();
-	}
 
 	/**
 	 * Reloads the view's data if the view is currently shown.
@@ -397,191 +289,7 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 				i++;
 			}
 		}
-
-		if (currentlyShownWeek != null) {
-			DateTime monday = DateTimeUtil.stringToDateTime(currentlyShownWeek.getStart());
-			DateTime tuesday = monday.plusDays(1);
-			DateTime wednesday = tuesday.plusDays(1);
-			DateTime thursday = wednesday.plusDays(1);
-			DateTime friday = thursday.plusDays(1);
-			DateTime saturday = friday.plusDays(1);
-			DateTime sunday = saturday.plusDays(1);
-			// set dates
-			showActualDates(monday, tuesday, wednesday, thursday, friday, saturday, sunday);
-			// highlight current day (if it is visible)
-			// and reset the highlighting for the other days
-			refreshRowHighlighting(monday, tuesday, wednesday, thursday, friday, saturday, sunday);
-			// display times
-			showTimes(monday, tuesday, wednesday, thursday, friday, saturday, sunday);
-		}
-	}
-
-	private void refreshRowHighlighting(DateTime monday, DateTime tuesday, DateTime wednesday, DateTime thursday,
-		DateTime friday, DateTime saturday, DateTime sunday) {
-		DateTime today = DateTimeUtil.getCurrentDateTime();
-		mondayRow.setBackgroundResource(today.isSameDayAs(monday) ? R.drawable.table_row_highlighting
-			: R.drawable.table_row);
-		tuesdayRow.setBackgroundResource(today.isSameDayAs(tuesday) ? R.drawable.table_row_highlighting : 0);
-		wednesdayRow.setBackgroundResource(today.isSameDayAs(wednesday) ? R.drawable.table_row_highlighting
-			: R.drawable.table_row);
-		thursdayRow.setBackgroundResource(today.isSameDayAs(thursday) ? R.drawable.table_row_highlighting : 0);
-		fridayRow.setBackgroundResource(today.isSameDayAs(friday) ? R.drawable.table_row_highlighting
-			: R.drawable.table_row);
-		saturdayRow.setBackgroundResource(today.isSameDayAs(saturday) ? R.drawable.table_row_highlighting : 0);
-		sundayRow.setBackgroundResource(today.isSameDayAs(sunday) ? R.drawable.table_row_highlighting
-			: R.drawable.table_row);
-	}
-
-	private void showActualDates(DateTime monday, DateTime tuesday, DateTime wednesday, DateTime thursday,
-		DateTime friday, DateTime saturday, DateTime sunday) {
-		topLeftCorner.setText("W " + thursday.getWeekIndex(DateTimeUtil.getBeginOfFirstWeekFor(thursday.getYear())));
-		mondayLabel.setText(getString(R.string.monday) + getString(R.string.onespace)
-			+ monday.format(getString(R.string.shortDateFormat)));
-		tuesdayLabel.setText(getString(R.string.tuesday) + getString(R.string.onespace)
-			+ tuesday.format(getString(R.string.shortDateFormat)));
-		wednesdayLabel.setText(getString(R.string.wednesday) + getString(R.string.onespace)
-			+ wednesday.format(getString(R.string.shortDateFormat)));
-		thursdayLabel.setText(getString(R.string.thursday) + getString(R.string.onespace)
-			+ thursday.format(getString(R.string.shortDateFormat)));
-		fridayLabel.setText(getString(R.string.friday) + getString(R.string.onespace)
-			+ friday.format(getString(R.string.shortDateFormat)));
-		saturdayLabel.setText(getString(R.string.saturday) + getString(R.string.onespace)
-			+ saturday.format(getString(R.string.shortDateFormat)));
-		sundayLabel.setText(getString(R.string.sunday) + getString(R.string.onespace)
-			+ sunday.format(getString(R.string.shortDateFormat)));
-	}
-
-	private void showTimes(DateTime monday, DateTime tuesday, DateTime wednesday, DateTime thursday, DateTime friday,
-		DateTime saturday, DateTime sunday) {
-		if (currentlyShownWeek != null) {
-			TimeSum flexiBalance = null;
-			boolean hasRealData = !(currentlyShownWeek instanceof WeekPlaceholder);
-			if (hasRealData && preferences.getBoolean(Key.ENABLE_FLEXI_TIME.getName(), false)) {
-				flexiBalance = timerManager.getFlexiBalanceAtWeekStart(DateTimeUtil.stringToDateTime(currentlyShownWeek
-					.getStart()));
-			}
-			boolean earlierEventsExist = (dao.getLastEventBefore(monday.getStartOfDay()) != null);
-			boolean showFlexiTimes = hasRealData || earlierEventsExist;
-			List<Event> events = fetchEventsForDay(monday);
-			flexiBalance = showTimesForSingleDay(monday, events, flexiBalance, mondayIn, mondayOut, mondayWorked,
-				mondayFlexi, showFlexiTimes);
-			events = fetchEventsForDay(tuesday);
-			flexiBalance = showTimesForSingleDay(tuesday, events, flexiBalance, tuesdayIn, tuesdayOut, tuesdayWorked,
-				tuesdayFlexi, showFlexiTimes);
-			events = fetchEventsForDay(wednesday);
-			flexiBalance = showTimesForSingleDay(wednesday, events, flexiBalance, wednesdayIn, wednesdayOut,
-				wednesdayWorked, wednesdayFlexi, showFlexiTimes);
-			events = fetchEventsForDay(thursday);
-			flexiBalance = showTimesForSingleDay(thursday, events, flexiBalance, thursdayIn, thursdayOut,
-				thursdayWorked, thursdayFlexi, showFlexiTimes);
-			events = fetchEventsForDay(friday);
-			flexiBalance = showTimesForSingleDay(friday, events, flexiBalance, fridayIn, fridayOut, fridayWorked,
-				fridayFlexi, showFlexiTimes);
-			events = fetchEventsForDay(saturday);
-			flexiBalance = showTimesForSingleDay(saturday, events, flexiBalance, saturdayIn, saturdayOut,
-				saturdayWorked, saturdayFlexi, showFlexiTimes);
-			events = fetchEventsForDay(sunday);
-			flexiBalance = showTimesForSingleDay(sunday, events, flexiBalance, sundayIn, sundayOut, sundayWorked,
-				sundayFlexi, showFlexiTimes);
-
-			TimeSum amountWorked = timerManager.calculateTimeSum(DateTimeUtil.getWeekStart(DateTimeUtil
-				.stringToDateTime(currentlyShownWeek.getStart())), PeriodEnum.WEEK);
-			showSummaryLine(amountWorked, flexiBalance, showFlexiTimes && DateTimeUtil.isInPast(monday
-				.getStartOfDay()));
-		}
-	}
-
-	private List<Event> fetchEventsForDay(DateTime day) {
-		Logger.debug("fetchEventsForDay: {}", DateTimeUtil.dateTimeToDateString(day));
-		List<Event> ret = dao.getEventsOnDay(day);
-		DateTime now = DateTimeUtil.getCurrentDateTime();
-		Event lastEventBeforeNow = dao.getLastEventBefore(now);
-		if (day.isSameDayAs(now) && TimerManager.isClockInEvent(lastEventBeforeNow)) {
-			// currently clocked in: add clock-out event "NOW"
-			ret.add(timerManager.createClockOutNowEvent());
-		}
-		return ret;
-	}
-
-	private TimeSum showTimesForSingleDay(DateTime day, List<Event> events, TimeSum flexiBalanceAtDayStart,
-		TextView in, TextView out, TextView worked, TextView flexi, boolean showFlexiTimes) {
-
-		DayLine dayLine = timeCalculator.calulateOneDay(day, events);
-
-		WeekDayEnum weekDay = WeekDayEnum.getByValue(day.getWeekDay());
-		boolean isWorkDay = timerManager.isWorkDay(weekDay);
-		boolean isTodayOrEarlier = DateTimeUtil.isInPast(day.getStartOfDay());
-		boolean containsEventsForDay = containsEventsForDay(events, day);
-		boolean weekEndWithoutEvents = !isWorkDay && !containsEventsForDay;
-		// correct result by previous flexi time sum
-		dayLine.getTimeFlexi().addOrSubstract(flexiBalanceAtDayStart);
-
-		in.setText(formatTime(dayLine.getTimeIn()));
-		if (isCurrentMinute(dayLine.getTimeOut()) && timerManager.isTracking()) {
-			out.setText("NOW");
-		} else {
-			out.setText(formatTime(dayLine.getTimeOut()));
-		}
-		if (weekEndWithoutEvents) {
-			worked.setText("");
-		} else if (isWorkDay && isTodayOrEarlier) {
-			worked.setText(formatSum(dayLine.getTimeWorked(), null));
-		} else {
-			worked.setText(formatSum(dayLine.getTimeWorked(), ""));
-		}
-		if (!showFlexiTimes || weekEndWithoutEvents || !preferences.getBoolean(Key.ENABLE_FLEXI_TIME.getName(),
-			false)) {
-			flexi.setText("");
-		} else if (isWorkDay && isTodayOrEarlier) {
-			flexi.setText(formatSum(dayLine.getTimeFlexi(), null));
-		} else if (containsEventsForDay) {
-			flexi.setText(formatSum(dayLine.getTimeFlexi(), ""));
-		} else {
-			flexi.setText("");
-		}
-
-		return dayLine.getTimeFlexi();
-	}
-
-	private boolean containsEventsForDay(List<Event> events, DateTime day) {
-		for (Event event : events) {
-			if (DateTimeUtil.stringToDateTime(event.getTime()).isSameDayAs(day)) {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	private boolean isCurrentMinute(DateTime dateTime) {
-		if (dateTime == null) {
-			return false;
-		}
-		DateTime now = DateTimeUtil.getCurrentDateTime();
-		return now.getYear().equals(dateTime.getYear())
-			&& now.getMonth().equals(dateTime.getMonth())
-			&& now.getDay().equals(dateTime.getDay())
-			&& now.getHour().equals(dateTime.getHour())
-			&& now.getMinute().equals(dateTime.getMinute());
-	}
-
-	private String formatTime(DateTime time) {
-		return time == null ? "" : DateTimeUtil.dateTimeToHourMinuteString(time);
-	}
-
-	private String formatSum(TimeSum sum, String valueForZero) {
-		if (sum != null && sum.getAsMinutes() == 0 && valueForZero != null) {
-			return valueForZero;
-		}
-		return sum == null ? "" : sum.toString();
-	}
-
-	private void showSummaryLine(TimeSum amountWorked, TimeSum flexiBalance, boolean showFlexiTimes) {
-		totalWorked.setText(amountWorked.toString());
-		if (flexiBalance != null && showFlexiTimes) {
-			totalFlexi.setText(flexiBalance.toString());
-		} else {
-			totalFlexi.setText("");
-		}
+		weekFragment.refreshView();
 	}
 
 	private void setupTasksAdapter() {
@@ -592,68 +300,10 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 	}
 
 	private void findAllViewsById() {
-		weekTable = findViewById(R.id.week_table);
-		titleRow = findViewById(R.id.titleRow);
-		topLeftCorner = findViewById(R.id.topLeftCorner);
-		inLabel = findViewById(R.id.inLabel);
-		outLabel = findViewById(R.id.outLabel);
-		workedLabel = findViewById(R.id.workedLabel);
-		flexiLabel = findViewById(R.id.flexiLabel);
-		mondayRow = findViewById(R.id.mondayRow);
-		mondayLabel = findViewById(R.id.mondayLabel);
-		mondayIn = findViewById(R.id.mondayIn);
-		mondayOut = findViewById(R.id.mondayOut);
-		mondayWorked = findViewById(R.id.mondayWorked);
-		mondayFlexi = findViewById(R.id.mondayFlexi);
-		tuesdayRow = findViewById(R.id.tuesdayRow);
-		tuesdayLabel = findViewById(R.id.tuesdayLabel);
-		tuesdayIn = findViewById(R.id.tuesdayIn);
-		tuesdayOut = findViewById(R.id.tuesdayOut);
-		tuesdayWorked = findViewById(R.id.tuesdayWorked);
-		tuesdayFlexi = findViewById(R.id.tuesdayFlexi);
-		wednesdayRow = findViewById(R.id.wednesdayRow);
-		wednesdayLabel = findViewById(R.id.wednesdayLabel);
-		wednesdayIn = findViewById(R.id.wednesdayIn);
-		wednesdayOut = findViewById(R.id.wednesdayOut);
-		wednesdayWorked = findViewById(R.id.wednesdayWorked);
-		wednesdayFlexi = findViewById(R.id.wednesdayFlexi);
-		thursdayRow = findViewById(R.id.thursdayRow);
-		thursdayLabel = findViewById(R.id.thursdayLabel);
-		thursdayIn = findViewById(R.id.thursdayIn);
-		thursdayOut = findViewById(R.id.thursdayOut);
-		thursdayWorked = findViewById(R.id.thursdayWorked);
-		thursdayFlexi = findViewById(R.id.thursdayFlexi);
-		fridayRow = findViewById(R.id.fridayRow);
-		fridayLabel = findViewById(R.id.fridayLabel);
-		fridayIn = findViewById(R.id.fridayIn);
-		fridayOut = findViewById(R.id.fridayOut);
-		fridayWorked = findViewById(R.id.fridayWorked);
-		fridayFlexi = findViewById(R.id.fridayFlexi);
-		saturdayRow = findViewById(R.id.saturdayRow);
-		saturdayLabel = findViewById(R.id.saturdayLabel);
-		saturdayIn = findViewById(R.id.saturdayIn);
-		saturdayOut = findViewById(R.id.saturdayOut);
-		saturdayWorked = findViewById(R.id.saturdayWorked);
-		saturdayFlexi = findViewById(R.id.saturdayFlexi);
-		sundayRow = findViewById(R.id.sundayRow);
-		sundayLabel = findViewById(R.id.sundayLabel);
-		sundayIn = findViewById(R.id.sundayIn);
-		sundayOut = findViewById(R.id.sundayOut);
-		sundayWorked = findViewById(R.id.sundayWorked);
-		sundayFlexi = findViewById(R.id.sundayFlexi);
-		totalRow = findViewById(R.id.totalRow);
-		totalLabel = findViewById(R.id.totalLabel);
-		totalIn = findViewById(R.id.totalIn);
-		totalOut = findViewById(R.id.totalOut);
-		totalWorked = findViewById(R.id.totalWorked);
-		totalFlexi = findViewById(R.id.totalFlexi);
-		previousWeekButton = findViewById(R.id.previous);
-		nextWeekButton = findViewById(R.id.next);
 		task = findViewById(R.id.task);
 		text = findViewById(R.id.text);
 		clockInButton = findViewById(R.id.clockInButton);
 		clockOutButton = findViewById(R.id.clockOutButton);
-		todayButton = findViewById(R.id.todayButton);
 	}
 
 	/**
@@ -736,10 +386,20 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 		}
 	}
 
+	@Override
+	public void onWeekTableClick() {
+		showEventList();
+	}
+
 	private void showEventList() {
 		Logger.debug("showing EventList");
 		Intent i = new Intent(this, EventListActivity.class);
-		i.putExtra(Constants.WEEK_START_EXTRA_KEY, currentlyShownWeek.getStart());
+		Week currentWeek = weekFragment.getWeek();
+		if(currentWeek == null) {
+			Logger.error("WeekFragment has no Week set");
+			return;
+		}
+		i.putExtra(Constants.WEEK_START_EXTRA_KEY, currentWeek.getStart());
 		startActivity(i);
 	}
 
@@ -834,12 +494,6 @@ public class WorkTimeTrackerActivity extends AppCompatActivity {
 			setupTasksAdapter();
 		}
 
-		String weekStart = DateTimeUtil.getWeekStartAsString(DateTimeUtil.getCurrentDateTime());
-		currentlyShownWeek = dao.getWeek(weekStart);
-		if (currentlyShownWeek == null) {
-			// don't insert a new week into the DB but only use a placeholder
-			currentlyShownWeek = new WeekPlaceholder(weekStart);
-		}
 		Basics.getInstance().safeCheckPersistentNotification();
 
 		refreshView();
