+++ app/src/main/res/layout/main.xml
@@ -14,41 +14,12 @@
         android:focusableInTouchMode="true"
         android:orientation="vertical" >
 
-        <LinearLayout
-            android:id="@+id/LinearLayout2"
-            android:layout_width="fill_parent"
+        <fragment
+            android:id="@+id/week_fragment"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="3dp"
-            android:orientation="horizontal" >
-
-            <Button
-                android:id="@+id/previous"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="@string/previous" />
-
-            <Button
-                android:id="@+id/todayButton"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="@string/today" />
-
-            <Button
-                android:id="@+id/next"
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:layout_weight="1"
-                android:text="@string/next" />
-        </LinearLayout>
-
-        <include
-            android:id="@+id/week_table"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="22dp"
-            layout="@layout/week_table" />
+            android:name="org.zephyrsoft.trackworktime.util.WeekFragment"
+            />
 
         <LinearLayout
             android:layout_width="fill_parent"
